refactor: restructure project into a workspace with client, shell, widgets, and application crates.
