+++ apps/yaml-editor/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "yaml-editor"
+description = "Dora Studio YAML pipeline editor"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[dependencies]
