+++ dora-studio-client/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "dora-studio-client"
+description = "Headless client library for Dora Studio: AI API, CLI tools, and OTLP telemetry"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[dependencies]
+serde.workspace = true
+serde_json.workspace = true
+
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+tokio.workspace = true
+reqwest.workspace = true
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+tokio = { version = "1", features = ["sync", "macros", "rt", "time"] }
+reqwest = { version = "0.12", default-features = false, features = ["json"] }
+wasm-bindgen-futures.workspace = true
