+++ dora-studio-client/src/lib.rs
@@ -1,9 +1,4 @@
-pub use makepad_widgets;
-
 pub mod api;
-pub mod app;
-pub mod chat;
-pub mod dataflow;
 
 // Tools module only available on native platforms (uses shell commands)
 #[cfg(not(target_arch = "wasm32"))]
@@ -12,7 +7,3 @@ pub mod tools;
 // OTLP telemetry client module only available on native platforms
 #[cfg(not(target_arch = "wasm32"))]
 pub mod otlp;
-
-// Traces panel module only available on native platforms
-#[cfg(not(target_arch = "wasm32"))]
-pub mod traces;
