+++ dora-studio-shell/src/lib.rs
@@ -1,0 +1,3 @@
+pub use makepad_widgets;
+
+pub mod app;
