+++ dora-studio-shell/src/main.rs
@@ -1,0 +1,3 @@
+fn main() {
+    dora_studio_shell::app::app_main()
+}
