--- src/chat/mod.rs
@@ -1,7 +1,0 @@
-use makepad_widgets::Cx;
-
-pub mod chat_screen;
-
-pub fn live_design(cx: &mut Cx) {
-    self::chat_screen::live_design(cx);
-}
