--- src/main.rs
@@ -1,3 +1,0 @@
-fn main() {
-    dora_studio::app::app_main()
-}
