--- tests/e2e/dataflow_lifecycle.rs
@@ -1,102 +1,0 @@
-//! E2E tests for complete dataflow lifecycle
-//!
-//! These tests require a running Dora coordinator and daemon.
-//! Run with: cargo test --test e2e -- --ignored
-
-use std::time::Duration;
-
-// ============================================================================
-// Full Lifecycle Tests
-// ============================================================================
-
-/// Test complete dataflow lifecycle: start -> monitor -> stop
-#[tokio::test]
-#[ignore] // Requires running Dora
-async fn test_full_dataflow_lifecycle() {
-    // TODO: Implement with real Dora connection
-    // let client = DoraClient::connect("127.0.0.1:53290").await.unwrap();
-    //
-    // // Start dataflow
-    // let uuid = client.start_dataflow("examples/simple.yaml").await.unwrap();
-    //
-    // // Verify running
-    // let flows = client.list_dataflows().await.unwrap();
-    // let flow = flows.iter().find(|f| f.uuid == uuid).unwrap();
-    // assert!(matches!(flow.status, DataflowStatus::Running));
-    //
-    // // Wait for nodes to initialize
-    // tokio::time::sleep(Duration::from_secs(2)).await;
-    //
-    // // Get node metrics
-    // let nodes = client.list_nodes(&uuid).await.unwrap();
-    // assert!(!nodes.is_empty());
-    //
-    // // Stop dataflow
-    // client.stop_dataflow(&uuid).await.unwrap();
-    //
-    // // Verify stopped
-    // tokio::time::sleep(Duration::from_secs(1)).await;
-    // let flows = client.list_dataflows().await.unwrap();
-    // let flow = flows.iter().find(|f| f.uuid == uuid);
-    // assert!(flow.is_none() || !matches!(flow.unwrap().status, DataflowStatus::Running));
-
-    todo!("Implement test_full_dataflow_lifecycle with real Dora")
-}
-
-/// Test dataflow reload (hot-reload)
-#[tokio::test]
-#[ignore] // Requires running Dora
-async fn test_dataflow_reload() {
-    // TODO: Test hot-reload functionality
-    // 1. Start dataflow
-    // 2. Modify YAML
-    // 3. Reload
-    // 4. Verify changes applied without restart
-
-    todo!("Implement test_dataflow_reload")
-}
-
-/// Test multiple concurrent dataflows
-#[tokio::test]
-#[ignore] // Requires running Dora
-async fn test_multiple_dataflows() {
-    // TODO: Test running multiple dataflows simultaneously
-    // 1. Start dataflow A
-    // 2. Start dataflow B
-    // 3. Verify both running
-    // 4. Stop A, verify B still running
-    // 5. Stop B
-
-    todo!("Implement test_multiple_dataflows")
-}
-
-// ============================================================================
-// Error Recovery Tests
-// ============================================================================
-
-/// Test recovery from node crash
-#[tokio::test]
-#[ignore] // Requires running Dora
-async fn test_node_crash_recovery() {
-    // TODO: Test behavior when a node crashes
-    // 1. Start dataflow
-    // 2. Simulate node crash
-    // 3. Verify status updated to Failed
-    // 4. Verify logs captured
-
-    todo!("Implement test_node_crash_recovery")
-}
-
-/// Test coordinator disconnect handling
-#[tokio::test]
-#[ignore] // Requires running Dora
-async fn test_coordinator_disconnect() {
-    // TODO: Test client behavior when coordinator disconnects
-    // 1. Connect to coordinator
-    // 2. Kill coordinator
-    // 3. Verify client handles gracefully
-    // 4. Restart coordinator
-    // 5. Verify client reconnects
-
-    todo!("Implement test_coordinator_disconnect")
-}
