--- tests/e2e/full_workflow.rs
@@ -1,141 +1,0 @@
-//! E2E tests for complete user workflows
-//!
-//! These tests simulate real user interactions across multiple apps.
-//! Run with: cargo test --test e2e -- --ignored
-
-use std::time::Duration;
-
-// ============================================================================
-// Workflow: Create and Monitor Dataflow
-// ============================================================================
-
-/// Complete workflow: create YAML -> start -> view logs -> view metrics -> stop
-#[tokio::test]
-#[ignore] // Requires running Dora
-async fn test_workflow_create_and_monitor() {
-    // TODO: Implement full workflow test
-    //
-    // 1. YAML Editor: Create new dataflow YAML
-    //    - Open editor
-    //    - Type YAML content
-    //    - Verify graph preview updates
-    //    - Save to file
-    //
-    // 2. Dataflow Manager: Start dataflow
-    //    - Click start button
-    //    - Select saved YAML file
-    //    - Wait for running status
-    //
-    // 3. Log Viewer: Monitor logs
-    //    - Switch to Log Viewer
-    //    - Verify logs streaming
-    //    - Apply level filter
-    //    - Search for specific message
-    //
-    // 4. Telemetry Dashboard: View metrics
-    //    - Switch to Telemetry Dashboard
-    //    - Verify CPU/Memory charts updating
-    //    - Check golden signals
-    //
-    // 5. Dataflow Manager: Stop dataflow
-    //    - Switch back to Dataflow Manager
-    //    - Select running dataflow
-    //    - Click stop
-    //    - Verify stopped status
-
-    todo!("Implement test_workflow_create_and_monitor")
-}
-
-// ============================================================================
-// Workflow: Debug Performance Issue
-// ============================================================================
-
-/// Workflow: identify and debug performance bottleneck
-#[tokio::test]
-#[ignore] // Requires running Dora
-async fn test_workflow_debug_performance() {
-    // TODO: Implement performance debugging workflow
-    //
-    // 1. Start a CPU-intensive dataflow
-    //
-    // 2. Telemetry Dashboard: Identify bottleneck
-    //    - View CPU chart
-    //    - Identify node with high CPU
-    //    - Check latency percentiles
-    //
-    // 3. Log Viewer: Find related errors
-    //    - Filter by bottleneck node
-    //    - Search for warnings/errors
-    //
-    // 4. AI Agent: Ask for analysis
-    //    - "What's causing the high CPU?"
-    //    - Verify tool calls made
-    //    - Verify recommendations provided
-
-    todo!("Implement test_workflow_debug_performance")
-}
-
-// ============================================================================
-// Workflow: AI-Assisted Operations
-// ============================================================================
-
-/// Workflow: use AI agent to manage dataflows
-#[tokio::test]
-#[ignore] // Requires running Dora + LLM
-async fn test_workflow_ai_assisted() {
-    // TODO: Implement AI-assisted workflow
-    //
-    // 1. AI: "Start the camera pipeline"
-    //    - Verify start_dataflow tool called
-    //    - Verify dataflow started
-    //
-    // 2. AI: "Show me the logs from yolo node"
-    //    - Verify filter_logs tool called
-    //    - Verify results displayed
-    //
-    // 3. AI: "What's the p99 latency?"
-    //    - Verify query_metrics tool called
-    //    - Verify latency reported
-    //
-    // 4. AI: "Stop all dataflows"
-    //    - Verify list_dataflows + stop_dataflow called
-    //    - Verify all stopped
-
-    todo!("Implement test_workflow_ai_assisted")
-}
-
-// ============================================================================
-// Workflow: YAML Editing
-// ============================================================================
-
-/// Workflow: create complex dataflow via YAML editor
-#[tokio::test]
-#[ignore]
-async fn test_workflow_yaml_editing() {
-    // TODO: Implement YAML editing workflow
-    //
-    // 1. Open YAML Editor
-    //
-    // 2. Type initial nodes
-    //    - camera node
-    //    - Verify graph shows 1 node
-    //
-    // 3. Add detector node with connection
-    //    - Type detector node with input from camera
-    //    - Verify edge appears in graph
-    //
-    // 4. Introduce validation error
-    //    - Remove required field
-    //    - Verify error highlighting
-    //
-    // 5. Fix error
-    //    - Add field back
-    //    - Verify error cleared
-    //
-    // 6. Save and verify
-    //    - Save to file
-    //    - Reload
-    //    - Verify content matches
-
-    todo!("Implement test_workflow_yaml_editing")
-}
