--- tests/integration/app_tests.rs
@@ -1,153 +1,0 @@
-//! Integration tests for Mini-Apps
-//!
-//! Tests app state management and component interaction.
-
-use super::*;
-
-// ============================================================================
-// Dataflow Manager App Tests
-// ============================================================================
-
-mod dataflow_manager {
-    use super::*;
-
-    #[tokio::test]
-    async fn test_dataflow_list_loads_from_client() {
-        // TODO: Test loading dataflows from DoraClient
-        todo!("Implement test_dataflow_list_loads_from_client")
-    }
-
-    #[tokio::test]
-    async fn test_start_action_calls_client() {
-        // TODO: Test start button triggers client call
-        todo!("Implement test_start_action_calls_client")
-    }
-
-    #[tokio::test]
-    async fn test_stop_action_requires_selection() {
-        // TODO: Test stop requires selected dataflow
-        todo!("Implement test_stop_action_requires_selection")
-    }
-
-    #[tokio::test]
-    async fn test_auto_refresh_updates_list() {
-        // TODO: Test auto-refresh functionality
-        todo!("Implement test_auto_refresh_updates_list")
-    }
-}
-
-// ============================================================================
-// YAML Editor App Tests
-// ============================================================================
-
-mod yaml_editor {
-    use super::*;
-
-    #[tokio::test]
-    async fn test_yaml_change_updates_graph() {
-        // TODO: Test YAML edit triggers graph update
-        todo!("Implement test_yaml_change_updates_graph")
-    }
-
-    #[tokio::test]
-    async fn test_graph_layout_hierarchical() {
-        // TODO: Test graph uses hierarchical layout
-        todo!("Implement test_graph_layout_hierarchical")
-    }
-
-    #[tokio::test]
-    async fn test_validation_errors_shown() {
-        // TODO: Test validation errors display
-        todo!("Implement test_validation_errors_shown")
-    }
-
-    #[tokio::test]
-    async fn test_file_save_and_load() {
-        // TODO: Test file operations
-        todo!("Implement test_file_save_and_load")
-    }
-}
-
-// ============================================================================
-// Log Viewer App Tests
-// ============================================================================
-
-mod log_viewer {
-    use super::*;
-
-    #[tokio::test]
-    async fn test_log_filtering_combined() {
-        // TODO: Test combined level + node + text filters
-        todo!("Implement test_log_filtering_combined")
-    }
-
-    #[tokio::test]
-    async fn test_real_time_streaming() {
-        // TODO: Test real-time log updates
-        todo!("Implement test_real_time_streaming")
-    }
-
-    #[tokio::test]
-    async fn test_export_respects_filters() {
-        // TODO: Test export with active filters
-        todo!("Implement test_export_respects_filters")
-    }
-}
-
-// ============================================================================
-// Telemetry Dashboard App Tests
-// ============================================================================
-
-mod telemetry_dashboard {
-    use super::*;
-
-    #[tokio::test]
-    async fn test_chart_updates_on_time_range() {
-        // TODO: Test time range changes update charts
-        todo!("Implement test_chart_updates_on_time_range")
-    }
-
-    #[tokio::test]
-    async fn test_golden_signals_calculation() {
-        // TODO: Test golden signals computed correctly
-        todo!("Implement test_golden_signals_calculation")
-    }
-
-    #[tokio::test]
-    async fn test_trace_timeline_rendering() {
-        // TODO: Test trace waterfall view
-        todo!("Implement test_trace_timeline_rendering")
-    }
-}
-
-// ============================================================================
-// AI Agent Integration Tests
-// ============================================================================
-
-mod ai_agent {
-    use super::*;
-
-    #[tokio::test]
-    async fn test_agent_executes_tool_call() {
-        // TODO: Test tool call execution flow
-        todo!("Implement test_agent_executes_tool_call")
-    }
-
-    #[tokio::test]
-    async fn test_agent_handles_multiple_tools() {
-        // TODO: Test chained tool calls
-        todo!("Implement test_agent_handles_multiple_tools")
-    }
-
-    #[tokio::test]
-    async fn test_chatbar_sends_to_agent() {
-        // TODO: Test ChatBar -> AgentCoordinator flow
-        todo!("Implement test_chatbar_sends_to_agent")
-    }
-
-    #[tokio::test]
-    async fn test_context_includes_app_state() {
-        // TODO: Test context manager includes current app state
-        todo!("Implement test_context_includes_app_state")
-    }
-}
