--- tests/integration/client_tests.rs
@@ -1,103 +1,0 @@
-//! Integration tests for DoraClient
-//!
-//! Tests client interaction with mock coordinator.
-
-use super::*;
-
-// ============================================================================
-// Connection Tests
-// ============================================================================
-
-#[tokio::test]
-async fn test_client_with_mock_coordinator() {
-    // TODO: Implement when MockCoordinator is ready
-    // let mock = MockCoordinator::start().await;
-    // let client = DoraClient::connect(&mock.addr()).await.unwrap();
-    // assert!(client.is_connected());
-    // mock.shutdown().await;
-    todo!("Implement test_client_with_mock_coordinator")
-}
-
-#[tokio::test]
-async fn test_client_handles_malformed_response() {
-    // TODO: Test error handling for invalid JSON responses
-    // let mock = MockCoordinator::start().await;
-    // mock.set_response(b"invalid json{{{");
-    // let mut client = DoraClient::connect(&mock.addr()).await.unwrap();
-    // let result = client.list_dataflows().await;
-    // assert!(result.is_err());
-    todo!("Implement test_client_handles_malformed_response")
-}
-
-#[tokio::test]
-async fn test_client_reconnects_on_disconnect() {
-    // TODO: Test automatic reconnection
-    todo!("Implement test_client_reconnects_on_disconnect")
-}
-
-// ============================================================================
-// Dataflow Operation Tests
-// ============================================================================
-
-#[tokio::test]
-async fn test_list_dataflows_empty() {
-    // TODO: Test listing when no dataflows exist
-    // let mock = MockCoordinator::start().await;
-    // mock.set_dataflows(vec![]);
-    // let mut client = DoraClient::connect(&mock.addr()).await.unwrap();
-    // let flows = client.list_dataflows().await.unwrap();
-    // assert!(flows.is_empty());
-    todo!("Implement test_list_dataflows_empty")
-}
-
-#[tokio::test]
-async fn test_list_dataflows_multiple() {
-    // TODO: Test listing multiple dataflows
-    todo!("Implement test_list_dataflows_multiple")
-}
-
-#[tokio::test]
-async fn test_start_dataflow_success() {
-    // TODO: Test successful dataflow start
-    todo!("Implement test_start_dataflow_success")
-}
-
-#[tokio::test]
-async fn test_start_dataflow_file_not_found() {
-    // TODO: Test error when YAML file doesn't exist
-    todo!("Implement test_start_dataflow_file_not_found")
-}
-
-#[tokio::test]
-async fn test_stop_dataflow_success() {
-    // TODO: Test successful dataflow stop
-    todo!("Implement test_stop_dataflow_success")
-}
-
-#[tokio::test]
-async fn test_stop_dataflow_not_running() {
-    // TODO: Test error when stopping non-running dataflow
-    todo!("Implement test_stop_dataflow_not_running")
-}
-
-// ============================================================================
-// Log Subscription Tests
-// ============================================================================
-
-#[tokio::test]
-async fn test_log_subscription_receives_logs() {
-    // TODO: Test real-time log streaming
-    todo!("Implement test_log_subscription_receives_logs")
-}
-
-#[tokio::test]
-async fn test_log_subscription_handles_disconnect() {
-    // TODO: Test graceful handling of disconnect during streaming
-    todo!("Implement test_log_subscription_handles_disconnect")
-}
-
-#[tokio::test]
-async fn test_log_subscription_backpressure() {
-    // TODO: Test handling of high-volume log streams
-    todo!("Implement test_log_subscription_backpressure")
-}
