--- tests/integration/mod.rs
@@ -1,11 +1,0 @@
-//! Integration Tests
-//!
-//! Tests that verify multiple components working together.
-
-mod app_tests;
-mod client_tests;
-mod storage_tests;
-
-// Re-export test utilities
-pub use super::fixtures::*;
-pub use super::mocks::*;
