--- tests/integration/storage_tests.rs
@@ -1,79 +1,0 @@
-//! Integration tests for Storage (DataFusion + Parquet)
-//!
-//! Tests storage persistence and querying.
-
-use super::*;
-use tempfile::TempDir;
-
-// ============================================================================
-// Initialization Tests
-// ============================================================================
-
-#[tokio::test]
-async fn test_storage_creates_directories() {
-    // TODO: Verify storage creates required subdirectories
-    // let temp_dir = TempDir::new().unwrap();
-    // let storage = Storage::new(temp_dir.path()).await.unwrap();
-    // assert!(temp_dir.path().join("metrics").exists());
-    // assert!(temp_dir.path().join("logs").exists());
-    // assert!(temp_dir.path().join("spans").exists());
-    todo!("Implement test_storage_creates_directories")
-}
-
-// ============================================================================
-// Persistence Tests
-// ============================================================================
-
-#[tokio::test]
-async fn test_storage_persistence_across_restarts() {
-    // TODO: Test data survives storage restart
-    // let temp_dir = TempDir::new().unwrap();
-    // {
-    //     let storage = Storage::new(temp_dir.path()).await.unwrap();
-    //     storage.insert_metrics(&[test_metric()]).await.unwrap();
-    // }
-    // {
-    //     let storage = Storage::new(temp_dir.path()).await.unwrap();
-    //     let result = storage.query("SELECT COUNT(*) FROM metrics").await.unwrap();
-    //     assert!(result.num_rows() > 0);
-    // }
-    todo!("Implement test_storage_persistence_across_restarts")
-}
-
-// ============================================================================
-// Query Tests
-// ============================================================================
-
-#[tokio::test]
-async fn test_storage_time_range_query() {
-    // TODO: Test filtering by time range
-    todo!("Implement test_storage_time_range_query")
-}
-
-#[tokio::test]
-async fn test_storage_aggregation_query() {
-    // TODO: Test SQL aggregations (AVG, SUM, COUNT)
-    todo!("Implement test_storage_aggregation_query")
-}
-
-#[tokio::test]
-async fn test_storage_join_query() {
-    // TODO: Test joining metrics with logs
-    todo!("Implement test_storage_join_query")
-}
-
-// ============================================================================
-// Performance Tests
-// ============================================================================
-
-#[tokio::test]
-async fn test_storage_bulk_insert_performance() {
-    // TODO: Test inserting 100K records
-    todo!("Implement test_storage_bulk_insert_performance")
-}
-
-#[tokio::test]
-async fn test_storage_query_performance() {
-    // TODO: Test query performance on large dataset
-    todo!("Implement test_storage_query_performance")
-}
